Improve add-path latency and runtime stability in OpenLMlib
- tune SQLite for write-heavy local usage (`WAL`, `synchronous=NORMAL`, `busy_timeout`, `temp_store=MEMORY`, larger cache, higher `wal_autocheckpoint`)
- reduce FTS maintenance overhead by switching to `INSERT OR REPLACE` instead of delete+insert
- add runtime embedding prewarm to cut first-request warmup cost in long-lived sessions
- serialize FAISS add/search access with a runtime lock to avoid contention and jitter
- keep deferred flush behavior integrated with runtime state
- add storage test coverage for connection PRAGMA performance settings

+++ openlmlib/write_gate.py
@@ -163,7 +163,7 @@ class WriteGate:
                 adjusted -= 0.05
 
         adjusted = max(0.0, min(1.0, adjusted))
-        return round(adjusted, 3)
+        return float(round(float(adjusted), 3))
 
     @staticmethod
     def is_allowed(issues: List[ValidationIssue]) -> bool:
